change view.py codes and fix minor error
