+++ BSCapp/templates/app/page-adminDataInfo.html
@@ -117,7 +117,7 @@
 									<table class="table table-striped">
 										<thead>
 											<tr>
-                                                <th align="center" onclick=adminDataSortBase('user_id')>出售者</th>
+                                                <th align="center" >出售者</th>
 												<th align="center" onclick=adminDataSortBase('data_name')>数据名</th>
                                                 <th align="center" onclick=adminDataSortBase('data_info')>数据简介</th>
 												<th align="center" onclick=adminDataSortBase('timestamp')>上传时间</th>
