--- transactions/1517624898.32025.json
@@ -1,1 +1,0 @@
-{"in_coins": [], "out_coins": [], "timestamp": 1517624898.32025, "action": "login", "seller": "576b4c13-d63c-3eaf-95fa-546753fce14c", "buyer": "", "data_uuid": "", "credit": 0.0, "reviewer": ""}
