--- transactions/1517624907.89224.json
@@ -1,1 +1,0 @@
-{"in_coins": [], "out_coins": [], "timestamp": 1517624907.89224, "action": "login", "seller": "admin", "buyer": "", "data_uuid": "", "credit": 0.0, "reviewer": ""}
