--- transactions/1517625466.435003.json
@@ -1,1 +1,0 @@
-{"in_coins": [], "out_coins": [], "timestamp": "1517625466.435003", "action": "review_reject", "seller": "u1", "buyer": "", "data_uuid": "d3", "credit": 0.0, "reviewer": "admin"}
