--- transactions/1517626147.502456.json
@@ -1,1 +1,0 @@
-{"in_coins": [], "out_coins": [], "timestamp": 1517626147.502456, "action": "login", "seller": "u1", "buyer": "", "data_uuid": "", "credit": 0.0, "reviewer": ""}
