--- transactions/1517626174.00318.json
@@ -1,1 +1,0 @@
-{"in_coins": [], "out_coins": [], "timestamp": 1517626174.00318, "action": "login", "seller": "u1", "buyer": "", "data_uuid": "", "credit": 0.0, "reviewer": ""}
